Bug: Fixed bug with first zero upon typing.
The commit fixed the issue that occurs when the user starts entering numbers after page is loaded.
As the user entered numbers, the zero which loads with the page wouldn't be removed.
The initial if condition that checks for zero didn't work.
To solve this, after page load, the html for output div was set to number 0.